+++ custom_components/solarcharger/time.py
@@ -210,7 +210,7 @@ async def async_setup_entry(
         if len(times) > 0:
             coordinator.device_controls[
                 subentry.subentry_id
-            ].controller.charge_control.times = times
+            ].controller.charge_control.entities.times = times
             async_add_entities(
                 times.values(),
                 update_before_add=False,